solocheck-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.7
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
celery==5.3.6
redis==5.0.1
firebase-admin==6.3.0
sendgrid==6.11.0
python-multipart==0.0.6
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
